mengupdate: beberapa controller dan menghapus controller login, mengupdate model, menghapus migrasi users dan transaksi, menghapus model agen, menghapus view beranda

+++ database/migrations/2024_10_11_073322_add_category_id_to_items_table.php
@@ -12,8 +12,13 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('items', function (Blueprint $table) {
-            // Izinkan category_id null sementara
-            $table->foreignId('category_id')->nullable()->constrained()->onDelete('cascade');
+            // Tambahkan kolom 'category_id' hanya jika belum ada
+            if (!Schema::hasColumn('items', 'category_id')) {
+                $table->foreignId('category_id')
+                    ->nullable()
+                    ->constrained()
+                    ->onDelete('cascade');
+            }
         });
     }
 
@@ -23,8 +28,11 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('items', function (Blueprint $table) {
-            $table->dropForeign(['category_id']);
-            $table->dropColumn('category_id');
+            // Hapus foreign key dan kolom 'category_id' hanya jika kolom ini ada
+            if (Schema::hasColumn('items', 'category_id')) {
+                $table->dropForeign(['category_id']); // Hapus foreign key
+                $table->dropColumn('category_id');   // Hapus kolom
+            }
         });
     }
 };
